requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
PyJWT>=2.9
bcrypt>=4.1
python-dotenv>=1.0
